Change everything to an INt64 for future proofing.
Simplified the process of calculating just do newSteps = count -
lastSteps; as that is all we care about as the rest of the logic works
great :)

+++ StepCounter.Android/Database/StepEntry.cs
@@ -13,7 +13,7 @@ namespace MyStepCounterAndroid.Database
 
 		public int ID { get; set; }
 
-		public int Steps { get; set; }
+		public Int64 Steps { get; set; }
 
 		public DateTime Date { get; set; }
 	}
